Update basic-gluster.sh
Every time new glusterfs packages are built, there is a window between them being tagged (-testing) into the -test repo and eventually/later being released into the mirror repos. Usually this window is short, but not always.

At line 70 the -test repo is enabled, glusterfs-api-devel and dependencies are installed, and ganesha is built against a potentially newer release than what's released. At line 111 the rest of glusterfs is installed. This can fail due to mismatched dependencies if a newer glusterfs was installed from the -test repo.

Seems like a bit of a no-brainer to test with the same glusterfs that was used to build ganesha with.