Install git package on the reserved node.
The `checkpatch` job is failing due to missing `git` package.
Recently, the task execution was moved from pod to a reserved
node.

It was assumed that the node was hardened. In the commit, we are
ensuring the git and python packages are installed. These are
essential for executing the checks.

Signed-off-by: Pragadeeswaran Sathyanarayanan <pragadeeswaran.sathyanarayanan@ibm.com>