Fix checkpatch review publish
By default, GERRIT_PUBLISH was set to `false. The existing check
switching to `true` was removed during the movement to a reserved
node.

Adding the condition that enables publishing of patches. Also, the
fetched repository had only the top level branch. This caused an
issue with clang-format.

Now we are fetching a depth of 2 so that we can compare.

Signed-off-by: Pragadeeswaran Sathyanarayanan <pragadeeswaran.sathyanarayanan@ibm.com>